Move getContactSuggestions to MailEditor.
We need it this way because it imports things from MailEditor and it's
not allowed during tests in node.
